requirements.txt
azure-cognitiveservices-vision-face
azure-storage-blob
python-dotenv
Pillow
flask